+++ checks/domain/template-linux/input.nix
@@ -4,5 +4,4 @@ lib: lib.domain.templates.linux
   uuid = "2904419d-b283-4cfd-9f2c-7c3713ff809f";
   memory = { count = 6; unit = "GiB"; };
   storage_vol_path = /Storage/MyHD.qcow2;
-  mac_address = "52:54:00:02:04:06";
 }