+++ checks/network/bridge/input.nix
@@ -2,6 +2,5 @@ lib: lib.network.templates.bridge
 {
   name = "test-bridge";
   uuid = "d2102492-5797-429b-aa31-96b1b0d6f8e8";
-  mac_address = "52:54:00:01:02:03";
   subnet_byte = 74;
 }