+++ templates/domain/windows.nix
@@ -5,7 +5,6 @@ stuff@{ packages, guest-install, ... }:
 , uuid
 , memory ? { count = 4; unit = "GiB"; }
 , storage_vol_path
-, mac_address
 , install_vol_path ? null
 , nvram_path
 , virtio_net ? false
@@ -16,7 +15,7 @@ stuff@{ packages, guest-install, ... }:
 let
   base = import ./base.nix stuff
     {
-      inherit name uuid memory storage_vol_path mac_address install_vol_path virtio_net;
+      inherit name uuid memory storage_vol_path install_vol_path virtio_net;
     };
 in
 base //