+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 * fixed defect in Home Manager module where domains were defined before swtpm package was available
+* clean up virtdeclare & virtpurge error reporting
 
 ## [0.2.0] - 2024-02-08