+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 * fixed defect in Home Manager module where domains were defined before swtpm package was available
 * clean up virtdeclare & virtpurge error reporting
-* improve XML generation
-* provide templates for network and domains
+* improved XML generation
+* provided templates for network and domains
+* NixOS module: set libvirtd package to match libvirt
 
 ## [0.2.0] - 2024-02-08